+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [14.8.0] - 2023-03-02
+
 ### Changed
 
 - Adjust the tccpn stack heartbeat to improve cluster upgrades.
@@ -909,7 +911,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v14.7.1...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-operator/compare/v14.8.0...HEAD
+[14.8.0]: https://github.com/giantswarm/aws-operator/compare/v14.7.1...v14.8.0
 [14.7.1]: https://github.com/giantswarm/aws-operator/compare/v14.7.0...v14.7.1
 [14.7.0]: https://github.com/giantswarm/aws-operator/compare/v14.6.0...v14.7.0
 [14.6.0]: https://github.com/giantswarm/aws-operator/compare/v14.5.0...v14.6.0